pyproject.toml
[tool.poetry]
name = "httpx-cache"
version = "0.1.0"
description = "Caching for httpx."
authors = ["Ouail Bendidi <ouail.bendidi@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6.2,<4.0"
httpx = "^0.21.1"
attrs = "^21.2.0"
anyio = "^3.4.0"
msgpack = "^1.0.3"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}
mypy = "^0.910"
isort = "^5.9.3"
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pre-commit = "^2.15.0"
pytest-cases = "^3.6.5"
trio = "^0.19.0"
trio-typing = "^0.6.0"
respx = "^0.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
python_version = 3.8
exclude = "tests/"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["httpx_cache"]
known_third_party = ["httpx", "attrs", "msgpack", "anyio"]

[tool.pytest.ini_options]
addopts = "-ra -q -vv --cov=httpx_cache --cov-report=term --cov-report=xml"


[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
omit = ["**/tests/test_*.py"]
